# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Thrive API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /athlete/profile:
    get:
      operationId: getAthleteProfile
      tags: [athlete]
      summary: Get athlete profile
      responses:
        "200":
          description: Athlete profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteProfile"
    patch:
      operationId: updateAthleteProfile
      tags: [athlete]
      summary: Update athlete profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AthleteProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteProfile"
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List recent activities
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Log a new activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get a single activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "404":
          description: Not found
...
  /injuries:
    get:
      operationId: listInjuries
      tags: [injuries]
      summary: List athlete injury history
      responses:
        "200":
          description: List of injuries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Injury"
    post:
      operationId: createInjury
      tags: [injuries]
      summary: Log a new injury
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InjuryInput"
      responses:
        "201":
          description: Created injury
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Injury"
  /injuries/{id}:
    patch:
      operationId: updateInjury
      tags: [injuries]
      summary: Update injury status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InjuryInput"
      responses:
        "200":
          description: Updated injury
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Injury"
        "404":
          description: Not found
    delete:
      operationId: deleteInjury
      tags: [injuries]
      summary: Delete an injury record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
...
  schemas:
...
    AthleteProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: ["integer", "null"]
        weeklyMileageGoal:
          type: ["number", "null"]
        fitnessLevel:
          type: string
          enum: [beginner, intermediate, advanced, elite]
        primaryGoal:
          type: string
        restingHeartRate:
          type: ["integer", "null"]
        hrv:
          type: ["number", "null"]
        selectedCoach:
          type: ["string", "null"]
          enum: [avera, kai, nova, rex, null]
        userRole:
          type: ["string", "null"]
          enum: [athlete, coach, null]
        country:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        pr5k:
          type: ["string", "null"]
        pr10k:
          type: ["string", "null"]
        prHalf:
          type: ["string", "null"]
        prMarathon:
          type: ["string", "null"]
        healthNotes:
          type: ["string", "null"]
        injuryRiskScore:
          type: ["number", "null"]
        injuryRiskLevel:
          type: ["string", "null"]
          enum: [low, medium, high, critical, null]
        injuryRiskUpdatedAt:
          type: ["string", "null"]
          format: date-time
        contactMethod:
          type: ["string", "null"]
          enum: [email, phone, null]
        contactValue:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - fitnessLevel
        - primaryGoal
        - createdAt
    AthleteProfileInput:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        weeklyMileageGoal:
          type: number
        fitnessLevel:
          type: string
          enum: [beginner, intermediate, advanced, elite]
        primaryGoal:
          type: string
        restingHeartRate:
          type: integer
        hrv:
          type: number
        selectedCoach:
          type: string
          enum: [avera, kai, nova, rex]
        userRole:
          type: string
          enum: [athlete, coach]
        country:
          type: string
        state:
          type: string
        pr5k:
          type: string
        pr10k:
          type: string
        prHalf:
          type: string
        prMarathon:
          type: string
        healthNotes:
          type: string
        contactMethod:
          type: string
          enum: [email, phone]
        contactValue:
          type: string
    Injury:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: ["integer", "null"]
        injuryType:
          type: string
        bodyPart:
          type: string
        dateOccurred:
          type: string
          format: date
        dateRecovered:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [active, recovered]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - injuryType
        - bodyPart
        - dateOccurred
        - status
        - createdAt
    InjuryInput:
      type: object
      properties:
        injuryType:
          type: string
        bodyPart:
          type: string
        dateOccurred:
          type: string
          format: date
        dateRecovered:
          type: string
          format: date
        status:
          type: string
          enum: [active, recovered]
        notes:
          type: string
      required:
        - injuryType
        - bodyPart
        - dateOccurred
        - status
    ActivitySplit:
      type: object
      properties:
        split:
          type: integer
        distance:
          type: number
        elapsedTime:
          type: integer
        movingTime:
          type: integer
        elevationDifference:
          type: ["number", "null"]
        averageSpeed:
          type: number
        averageHeartrate:
          type: ["number", "null"]
        paceZone:
          type: ["integer", "null"]
      required:
        - split
        - distance
        - elapsedTime
        - movingTime
        - averageSpeed
    ActivityBestEffort:
      type: object
      properties:
        name:
          type: string
        elapsedTime:
          type: integer
        distance:
          type: number
      required:
        - name
        - elapsedTime
        - distance
    Activity:
      type: object
      properties:
        id:
          type: integer
        stravaActivityId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [easy_run, tempo_run, interval, long_run, race, cross_training, rest]
        distanceKm:
          type: ["number", "null"]
        durationMinutes:
          type: ["integer", "null"]
        avgHeartRate:
          type: ["integer", "null"]
        perceivedEffort:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        activityDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        movingTimeSeconds:
          type: ["integer", "null"]
        elapsedTimeSeconds:
          type: ["integer", "null"]
        elevationGainM:
          type: ["number", "null"]
        elevHighM:
          type: ["number", "null"]
        elevLowM:
          type: ["number", "null"]
        maxHeartRate:
          type: ["integer", "null"]
        avgCadence:
          type: ["number", "null"]
        avgSpeed:
          type: ["number", "null"]
        maxSpeed:
          type: ["number", "null"]
        calories:
          type: ["number", "null"]
        sufferScore:
          type: ["integer", "null"]
        avgWatts:
          type: ["number", "null"]
        avgTemp:
          type: ["integer", "null"]
        achievementCount:
          type: ["integer", "null"]
        prCount:
          type: ["integer", "null"]
        kudosCount:
          type: ["integer", "null"]
        commentCount:
          type: ["integer", "null"]
        athleteCount:
          type: ["integer", "null"]
        gearName:
          type: ["string", "null"]
        startDateLocal:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        mapPolyline:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        workoutType:
          type: ["integer", "null"]
        splits:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/ActivitySplit"
        bestEfforts:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/ActivityBestEffort"
      required:
        - id
        - type
        - activityDate
        - createdAt
    ActivityInput:
      type: object
      properties:
        type:
          type: string
          enum: [easy_run, tempo_run, interval, long_run, race, cross_training, rest]
        distanceKm:
          type: number
        durationMinutes:
          type: integer
        avgHeartRate:
          type: integer
        perceivedEffort:
          type: integer
        notes:
          type: string
        activityDate:
          type: string
          format: date
      required:
        - type
        - activityDate
...